Removed checkstyle changes
