--- inception-recommendation/src/main/java/de/tudarmstadt/ukp/inception/recommendation/project/ProjectRecommendersPanel.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<!--
-#Copyright 2017
-#Ubiquitous Knowledge Processing (UKP) Lab
-#Technische Universität Darmstadt
-#
-#Licensed under the Apache License, Version 2.0 (the "License");
-#you may not use this file except in compliance with the License.
-#You may obtain a copy of the License at
-# 
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-#Unless required by applicable law or agreed to in writing, software
-#distributed under the License is distributed on an "AS IS" BASIS,
-#WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#See the License for the specific language governing permissions and
-#limitations under the License.
--->
-<html xmlns:wicket="http://wicket.apache.org">
-<wicket:panel>
-  <div class="flex-content flex-v-container flex-gutter">
-    <div class="flex-content flex-h-container flex-gutter">
-      <div class="flex-sidebar flex-v-container flex-only-internal-gutter" wicket:id="recommenders" style="flex-basis: 30%;"></div>
-      <div class="flex-content flex-v-container flex-only-internal-gutter" wicket:id="recommenderEditor"></div>
-    </div>
-  </div>
-</wicket:panel>
-</html>
