--- inception-recommendation/src/main/java/de/tudarmstadt/ukp/inception/recommendation/project/RecommenderEditorPanel.html
@@ -1,128 +1,0 @@
-<!DOCTYPE html>
-<!--
-#Copyright 2017
-#Ubiquitous Knowledge Processing (UKP) Lab
-#Technische Universität Darmstadt
-#
-#Licensed under the Apache License, Version 2.0 (the "License");
-#you may not use this file except in compliance with the License.
-#You may obtain a copy of the License at
-# 
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-#Unless required by applicable law or agreed to in writing, software
-#distributed under the License is distributed on an "AS IS" BASIS,
-#WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#See the License for the specific language governing permissions and
-#limitations under the License.
--->
-<html xmlns:wicket="http://wicket.apache.org">
-<body>
-  <wicket:panel>
-    <form wicket:id="form" class="flex-content panel panel-default panel-flex">
-      <div class="panel-heading">
-        <h3 class="panel-title"><wicket:message key="details"/></h3>
-      </div>
-      <div class="scrolling panel-body">
-        <div class="col-sm-12 form-horizontal">
-          <div class="form-group" wicket:enclosure="name">
-            <label class="col-sm-3 control-label" wicket:for="name">
-              <wicket:label key="name"/>
-            </label>
-            <div class="col-sm-9">
-              <div class="input-group">
-                <input wicket:id="name" class="form-control" type="text"/>
-                <span class="input-group-addon">
-                  <input wicket:id="autoGenerateName" type="checkbox"/>
-                  <wicket:message key="autoGenerateName"/>
-                </span>
-              </div>
-            </div>
-          </div>
-          <div class="form-group" wicket:enclosure="enabled">
-            <div class="col-sm-offset-3 col-sm-9">
-              <div class="checkbox">
-                <label wicket:for="enabled">
-                  <input wicket:id="enabled" type="checkbox"/>
-                  <wicket:label key="enabled"/>
-                </label>
-              </div>
-            </div>  
-          </div>
-          <div class="form-group" wicket:enclosure="layer">
-            <label wicket:for="layer" class="col-sm-3 control-label">
-              <wicket:label key="layer"/>
-            </label>
-            <div class="col-sm-9">
-              <select wicket:id="layer" class="form-control" data-container="body"></select>
-            </div>
-          </div>
-          <div class="form-group" wicket:enclosure="feature">
-            <label wicket:for="feature" class="col-sm-3 control-label">
-              <wicket:label key="feature"/>
-            </label>
-            <div class="col-sm-9">
-              <select wicket:id="feature" class="form-control" data-container="body"></select>
-            </div>
-          </div>
-          <div class="form-group" wicket:enclosure="tool">
-            <label wicket:for="tool" class="col-sm-3 control-label">
-              <wicket:label key="tool"/>
-            </label>
-            <div class="col-sm-9">
-              <select wicket:id="tool" class="form-control" data-container="body"></select>
-            </div>
-          </div>
-          <div class="form-group" wicket:id="activationContainer">
-            <label class="col-sm-3 control-label">
-              <wicket:message key="activation"/>
-            </label>
-            <div class="col-sm-9">
-              <div class="form-group" wicket:enclosure="threshold">
-                <label wicket:for="threshold" class="col-sm-3 control-label">
-                  <wicket:label key="threshold"/>
-                </label>
-                <div class="col-sm-9">
-                  <input type="number" wicket:id="threshold" class="form-control"/>
-                </div>
-              </div>
-              <div class="form-group" wicket:enclosure="alwaysSelected">
-                <div class="col-sm-offset-3 col-sm-9">
-                  <div class="checkbox">
-                    <label wicket:for="alwaysSelected">
-                      <input wicket:id="alwaysSelected" type="checkbox">
-                      <wicket:label key="alwaysSelected"/>
-                    </label>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-          <div class="form-group" wicket:enclosure="maxRecommendations">
-            <label wicket:for="maxRecommendations" class="col-sm-3 control-label">
-              <wicket:label key="maxRecommendations"/>
-            </label>
-            <div class="col-sm-9">
-              <input type="number" wicket:id="maxRecommendations" class="form-control"/>
-            </div>
-          </div>
-          <div class="form-group" wicket:enclosure="statesForTraining">
-            <label wicket:for="statesForTraining" class="col-sm-3 control-label">
-              <wicket:label key="statesForTraining"/>
-            </label>
-            <div class="col-sm-9" wicket:id="statesForTraining"></div>
-          </div>
-          <div wicket:id="traitsContainer">
-            <div wicket:id="traits"></div>
-          </div>
-        </div>
-      </div>
-      <div class="panel-footer text-right">
-        <input type="button" wicket:id="delete" class="btn btn-danger pull-left" wicket:message="value:delete"></input>
-        <input type="submit" wicket:id="save" class="btn btn-primary" wicket:message="value:save"></input>
-        <input type="button" wicket:id="cancel" class="btn btn-default" wicket:message="value:cancel"></input>
-      </div>            
-    </form>
-  </wicket:panel>
-</body>
-</html>
