--- inception-recommendation/src/main/java/de/tudarmstadt/ukp/inception/recommendation/project/RecommenderListPanel.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<!--
-#Copyright 2017
-#Ubiquitous Knowledge Processing (UKP) Lab
-#Technische Universität Darmstadt
-#
-#Licensed under the Apache License, Version 2.0 (the "License");
-#you may not use this file except in compliance with the License.
-#You may obtain a copy of the License at
-# 
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-#Unless required by applicable law or agreed to in writing, software
-#distributed under the License is distributed on an "AS IS" BASIS,
-#WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#See the License for the specific language governing permissions and
-#limitations under the License.
--->
-<html xmlns:wicket="http://wicket.apache.org">
-<body>
-  <wicket:panel>
-    <div class="flex-content panel panel-default panel-flex">
-      <div class="panel-heading">
-        <h3 class="panel-title"><wicket:message key="recommenders"/></h3>
-      </div>
-      <div class="panel-body fit-child-snug">
-        <select class="form-control" wicket:id="recommenders"></select>
-      </div>
-      <div class="panel-footer text-right">
-        <input wicket:id="create" type="button" class="btn btn-primary" wicket:message="value:create"/>
-      </div>            
-    </div>
-  </wicket:panel>
-</body>
-</html>
