--- inception-recommendation/src/main/java/de/tudarmstadt/ukp/inception/recommendation/sidebar/LearningCurveChartPanel.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<!--
-#Copyright 2018
-#Ubiquitous Knowledge Processing (UKP) Lab
-#Technische Universität Darmstadt
-#
-#Licensed under the Apache License, Version 2.0 (the "License");
-#you may not use this file except in compliance with the License.
-#You may obtain a copy of the License at
-# 
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-#Unless required by applicable law or agreed to in writing, software
-#distributed under the License is distributed on an "AS IS" BASIS,
-#WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#See the License for the specific language governing permissions and
-#limitations under the License.
--->
-<html xmlns:wicket="http://wicket.apache.org">
-<wicket:panel>
-  <div class="panel panel-default panel-small">
-    <div class="panel-heading">
-      <h3 class="panel-title">
-        <wicket:message key="learningCurve" />
-      </h3>
-    </div>
-    <div class="panel-body">
-      <div wicket:id="chart-container"></div>
-    </div>
-  </div>
-</wicket:panel>
-</html>
