--- inception-recommendation/src/main/java/de/tudarmstadt/ukp/inception/recommendation/sidebar/RecommendationSidebar.html
@@ -1,62 +1,0 @@
-<!DOCTYPE html>
-<!--
-#Copyright 2017
-#Ubiquitous Knowledge Processing (UKP) Lab
-#Technische Universität Darmstadt
-#
-#Licensed under the Apache License, Version 2.0 (the "License");
-#you may not use this file except in compliance with the License.
-#You may obtain a copy of the License at
-# 
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-#Unless required by applicable law or agreed to in writing, software
-#distributed under the License is distributed on an "AS IS" BASIS,
-#WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#See the License for the specific language governing permissions and
-#limitations under the License.
--->
-<html xmlns:wicket="http://wicket.apache.org">
-<wicket:panel>
-  <div class="flex-content flex-v-container flex-gutter flex-only-internal-gutter">
-    <div class="flex-content panel panel-default panel-flex">
-      <div class="panel-heading">
-        <h3 class="panel-title"><wicket:message key="recommender"/></h3>
-      </div>
-      <div class="scrolling panel-body">
-        <form wicket:id="form" class="panel panel-default panel-small">
-          <div class="panel-heading">
-            <h3 class="panel-title"><wicket:message key="settings"/></h3>
-          </div>
-          <div class="panel-body">
-            <div class="col-sm-12">
-              <div class="form-horizontal">
-                <div class="form-group" wicket:enclosure="maxPredictions">
-                  <label class="col-sm-6 control-label"><wicket:message key="maxSuggestions"/></label>
-                  <div class="col-sm-6">
-                    <input wicket:id="maxPredictions" type="number" class="form-control"/>
-                  </div>
-                </div>
-                <div class="form-group" wicket:enclosure="showAllPredictions">
-                  <label class="col-sm-6 control-label"><wicket:message key="showAll"/></label>
-                  <div class="col-sm-6">
-                    <div class="checkbox">
-                      <label wicket:for="showAllPredictions">
-                        <input wicket:id="showAllPredictions" type="checkbox"/>
-                      </label>
-                    </div>
-                  </div>  
-                </div>
-              </div>
-            </div>
-          </div>
-          <div class="panel-footer text-right">
-            <input wicket:id="save" type="submit" class="btn btn-primary"  wicket:message="value:apply"/>
-          </div>
-        </form>
-        <div wicket:id="learningCurve" ></div>
-      </div>
-    </div>
-  </div>
-</wicket:panel>
-</html>
